backend/requirements.txt
# Core Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Web Framework (Alternative)
flask>=2.3.0
flask-cors>=4.0.0

# Database
psycopg[binary,pool]>=3.1.0
psycopg2-binary>=2.9.0

# Environment & Config
python-dotenv>=1.0.0

# Machine Learning & Data Science
numpy>=1.24.0
scikit-learn>=1.3.0
umap-learn>=0.5.0
pandas>=2.0.0

# Vector Database & Embeddings
pgvector>=0.2.0

# Google Cloud AI
vertexai>=1.38.0

# Data Processing
json5>=0.9.0
typing-extensions>=4.8.0

# Logging & Monitoring
structlog>=23.2.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development (optional)
black>=23.0.0
flake8>=6.0.0

mcp
lightrag-hku
